Server tunneling code appears to work with patched TI-RPC library. Issues with RPC program registration remain.
This code compiles.  Many of the problems of the server-side
tunneling mechanism were solved in sstub_cntrl.c and sstub_conn.c.
Additionally, readn and writen were moved to a new common file,
socket_trans.h.  Potemkin was updated to use the hoststub interface.
Connection establishment appears to work flawlessly now, however
this uses the broken svc_run() in libtirpc and needs to be changed
to its own handler routine.