Fixed thread exiting bug
Threads were exiting from a misinterpretation of a failure code from
"request statistics" calls.  The call returns failure if no search
is currently running to request statistics about.  Unfortunately,
there is no easy way for adiskd to indicate to the storagestub library
that this specifically took place and another error did not occur, since
it returns a pointer to stats or NULL.  A new error code was added
to symbolize this opcode-specific error, but for now all errors coming from
adiskd use this new error code until the storagestub/adiskd interface
is changed.