Rename val and base_val to local_val and global_val
There will need to be some mechanism for not double-counting
state. When the client collects all the state from the servers,
it will compose the states together and then send them back to
each server. The client can do this whenever it wants, as often
as it wants. I think keeping these two variables distinct will
allow this to work. Somehow the value is latched. But is zero
a valid reset value?