Allow cache_lookup to work even when adding cache entries is disabled.
There are two dctl toggles, use_cache_table and cache_table.
When use_cache_table is set we will perform lookups in the ocache
When cache_table is set we will add new entries to the cache.

The same is valid of use_cache_oattr and cache_oattr, however these
do depend on the use_cache_table/cache_table settings to be enabled.
We use the cache hit from the ocache to know which attributes we
should read.