Some improvements on the ocache bugfix.
The missing pointer dereference was in fact the caller of the
cache_add_initial_attrs function. So move the dereference one level
back up as the rest of the code also assumes that lf_obj_handle is
really *obj_data_t.

Also make sure we terminate the sql transaction when we do not want
to replace existing initial attributes in the cache.