Remove slowdown.
It doesn't actually slow down anything, but starts a process that uses
a lot of CPU time. This doesn't really do anything on a multi-core or
hyperthreaded system, it relied on a hardcoded location of the slowdown
binary (/home/diamond/bin/slowdown), and used x86 specific assembly so
it failed to build on other CPU architectures.