Remove various heuristics that disable the attribute cache
The code is trying to see if we are IO-bound or CPU-bound and when
we are IO-bound it disables attribute caching. However a correctly
working attribute cache will avoid execution of expensive filters
and make the system appear to be IO-bound (resulting in disabling
of the cache). A face detection search across a relatively small
set of images went from 5-6 seconds down to 2-3 seconds if we do
not disable the attribute cache.