Move credit accounting to server
The credit model was broken. Since there are so many
queues, including the miniRPC queues, the server would
often send many more objects than were desired. This
was because the existing credit flow control model did
not take into account all these intermediate queues.

Replace this with a simple system that starts with some
number of credits on the server that is replenished
when the client actually consumes an object. This pushes
the accounting more to the endpoints and works much better.

The client will still crash if connecting to more than
OBJ_RING_SIZE/DEFAULT_QUEUE_LEN servers.