Add a proxy data retriever source.
Can be used to process a remote data retriever source from multiple
adiskd processes.

If you specify a scope url like,
    /proxy/1of2/server/scope/?options

it will get forwarded as,
    http://server/scope/?options

Returned object urls are filtered so that only every other object is
returned and are rewritten as absolute urls which point at the remote
data retriever.