Send back credits as soon as object are consumed on the client.
We used to only send credit updates 5 times per second, and combined
with the 10 object deep queue we would never send more than 50 objects
per server per second. By sending back as soon as an object is consumed
we have a in-flight/waiting to dequeue window of 10 objects per server.