Disable workahead.
If there are few or no filters and we queue a lot for the client it
makes the server waste a lot of computation on possibly unnecessary
results and causes duplicate work and more randomized disk accesses
in case the client is interested. If there is significant filter
execution and a lot of objects are dropped we are not as likely to
become stalled where workahead would kick in.