adiskd: Create a unique temporary directory for each search
When we are forking for a new search, have the child set TMPDIR to a
unique temporary directory created for the search.  When the child dies,
have the parent rm -rf the directory.  This allows filters to arrange that
their temporary data will be cleaned up regardless of how they terminate.