Normalize mtime on constructed opendiamond/__init__.py
The mtime of a Python source file is encoded in .pyc and .pyo files.  We
rewrite opendiamond/__init__.py on every build, changing the mtime, which
therefore causes RPM file conflicts between its .pyc and .pyo files when
installing both the 32-bit and 64-bit packages.  Fix this by setting the
mtime of the generated source file to the mtime of the file it clobbers.