diamondd: IPv6 listen fixes
getaddrinfo() sometimes returns the IPv4 address first and sometimes the
IPv6, causing us not to listen on both protocols depending on system
configuration.  Listen on all addresses returned by getaddrinfo, not just
the first successful one.  Also force v6 sockets not to bind to the IPv4
port to prevent conflicts.