Drop automatic manifest generation for Python-implemented filters
The support infrastructure for Python-implemented filters, unlike for
other kinds of filters, went to great lengths to ensure that the filter
author could declare the entire contents of the search manifest within
Python rather than having to construct the manifest by hand.  This made
sense with the non-human-readable opendiamond-manifest.txt format, but
makes less sense with XML.  It also required a complete engine for
converting Pythonic declarations into equivalent XML syntax; this posed
a maintainance problem as features were added to the XML manifest and did
not really add value.