opendiamond.filter: Move egg loading to a helper method
Rename Filter.blob_format to Filter.blob_is_zip and change it to a
boolean.  Add Filter.load_egg() method which can load an egg from the
entire blob or from an individual member if the blob is a Zip file.
Have Filter.load_egg() perform the actual import if requested.