diamondd: Remove search ID/data from get_object reply
In the past, search ID was used for relating objects to searches
because the same control and blast channels could be used for more than
one search. When the second search started, it was necessary to
determine whether an object on the blast channel was a remnant of the
first search or was produced by the second one. Now, a control and blast
channels pair is unique for each search, and search ID does not need to
be used in order to relate objects to searches.

Also, remove data field since it is not used anymore. Object data is now
stored in the attributes list with the zero-length string key.

Change get_object RPC (from 1 to 2).