diamondd: Fix blast channel deadlock when client does not pipeline
Transmitting and receiving on an RPCConnection acquire the connection
lock.  If worker thread 1 receives a blast channel request and goes to
respond to it, the transmit call has to contend for the lock with other
workers attempting to receive.  If worker thread 2 is blocked on receive
because no requests are pending on the channel, thread 1 will thus be
unable to transmit.  This can cause deadlocks when the number of blast
requests pipelined by the client is less than then number of worker
threads.

Fix this by holding the RPCConnection lock over an entire dispatch() call
rather than only in _receive() and _reply().  This removes the need to
reacquire the lock before transmitting.  It also serializes request
handlers on a particular RPCConnection, but since we're not trying to
support out-of-order replies in the wire protocol anymore, this might be
considered a feature.