diamondd: Convert result and attribute caches to Murmur hash
diamondd spends the majority of its time computing hashes for the result
and attribute caches.  MD5 is slow and is no longer resistant to malicious
collisions, and any secure hash function would be even slower.  Murmur is
five times as fast as MD5 and spreads well, but is not resistant to
malice.  This is fine for the result cache, for which a collision only
results in a cache miss.  However, the attribute cache -- for which we
perform most of the hashing -- is content-addressable.  For now, we will
live with the possibility of malicious collisions in the attribute cache;
if this becomes a problem, we can hash attribute data with a secret key.
(We would need to store the key somewhere inaccessible to a malicious
filter.)