diamondd: Fix incorrect cache entries when an attribute is overwritten
Consider a search in which filter A produces an attribute k -> v1 and
filter B then overwrites it with an attribute k -> v2.  If we decided to
cache_output on filter A, we would incorrectly produce attribute cache
entries as murmur(v1) -> v2.  If a subsequent search included filter A
but not filter B, the wrong attribute value v2 would then be returned
for k.  Fix this by not storing an attribute cache entry if the associated
value is no longer available.