setup.py:data_files works inconsistent
When building with `setup.py install` the files get installed in
the python package directory. But when we install as a 'flat' egg
using `setup.py install --single-version-externally-managed` they
are installed under `${prefix}`.

Moving `bundle.xsd` into the opendiamond python package directory
so that we can use package_data to install the file which does seem
to install to the same place consistently.