Don't cache opendiamond source
Since we only refresh the cache in setup and share it between all jobs
on the same branch, when a build is retried later it build against
whichever version of opendiamond happened to be build last instead of
the checked out tree. So we shouldn't try to cache the source.